requirements.txt
numpy>=2.2.0
opencv-python>=4.10.0.84
pillow>=11.0.0
pycuda>=2024.1.2
torch>=2.5.1
tqdm>=4.67.1
imageio>=2.36.1
torchvision>=0.8.0
spandrel>=0.4.0
nuitka>=2.5.8